refactor: 회원가입 컨트롤러 어노테이션 기반으로 변경해도 동작하는 것 확인
