+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -2,11 +2,14 @@ package com.techcourse.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
 
-public class RegisterViewController implements Controller {
+@Controller
+public class RegisterViewController {
 
-    @Override
+    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
     public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
         return "/register.jsp";
     }
