+++ mvc/src/main/java/nextstep/mvc/handler/adapter/AnnotationHandlerAdapter.java
@@ -3,6 +3,7 @@ package nextstep.mvc.handler.adapter;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.handler.HandlerExecution;
+import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 
 public class AnnotationHandlerAdapter implements HandlerAdapter {
@@ -15,6 +16,15 @@ public class AnnotationHandlerAdapter implements HandlerAdapter {
 	@Override
 	public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
 		HandlerExecution handlerExecution = (HandlerExecution)handler;
-		return handlerExecution.handle(request, response);
+		Object result = handlerExecution.handle(request, response);
+		return handleResult(result);
+	}
+
+	private ModelAndView handleResult(Object result) {
+		if (result instanceof ModelAndView) {
+			return (ModelAndView)result;
+		}
+		String viewName = (String)result;
+		return new ModelAndView(new JspView(viewName));
 	}
 }
